全部改为float并增加checkCudaErrors函数，试图定位问题

+++ src/Flow.h
@@ -6,8 +6,8 @@
 struct FlowData {
 	unsigned int src;
 	unsigned int dest;
-	double flow;
-	double dist;
+	float flow;
+	float dist;
 };
 
 namespace Flow {
