+++ src/utils.h
@@ -5,13 +5,13 @@
 #include <iostream>
 #include "DataConfig.h"
 
-const double PI = acos(-1.0);
+const float PI = acos(-1.0);
 
 // Random number between [0, 1] with normal distribution
-double BoxMullerRandom();
+float BoxMullerRandom();
 
 // Ramdom number beteween [0, 1]
-double random01();
+float random01();
 
 void parseArgs(int argc, char* argv[]);
